Split Remote to BarRemote
